core(style): readability fixes
